requirements.txt
# Slack Integration
slack-bolt==1.16.1
slack-sdk==3.19.5

# Google Cloud Integration
google-cloud-monitoring==2.14.2
google-cloud-recommender==2.11.2
google-cloud-container==2.17.4
google-auth==2.16.2

# Kubernetes Integration
kubernetes==26.1.0

# Jira Integration
jira==3.5.1

# Natural Language Processing
google-cloud-aiplatform==1.25.0
langchain==0.0.267
langchain-google-genai==0.0.5

# Web Framework
flask==2.2.3
gunicorn==20.1.0

# Data Processing and Visualization
pandas==1.5.3
matplotlib==3.7.1
numpy==1.24.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0
requests==2.28.2

# Code Quality
pylint==2.17.0 